requirements.txt
allure-pytest==2.8.21
allure-python-commons==2.8.21
apipkg==1.5
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
brotlipy==0.7.0
certifi==2020.11.8
cffi @ file:///C:/ci/cffi_1605538148575/work
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
cryptography @ file:///C:/ci/cryptography_1605544557248/work
execnet==1.7.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iniconfig @ file:///tmp/build/80754af9/iniconfig_1602780191262/work
more-itertools @ file:///tmp/build/80754af9/more-itertools_1605111547926/work
packaging==20.4
pluggy==0.13.1
py @ file:///tmp/build/80754af9/py_1593446248552/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyMySQL @ file:///C:/ci/pymysql_1601577445570/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytest==6.1.2
pytest-forked @ file:///tmp/build/80754af9/pytest-forked_1597717567176/work
pytest-rerunfailures @ file:///tmp/build/80754af9/pytest-rerunfailures_1601657766215/work
pytest-xdist @ file:///tmp/build/80754af9/pytest-xdist_1598373491753/work
selenium @ file:///C:/ci/selenium_1594408106746/work
six @ file:///C:/ci/six_1605205426665/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
